pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "predictive-maintenance"
version = "0.1.0"
description = "Pro-level predictive maintenance system on NASA CMAPSS FD001–FD004"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "optuna>=3.4",
    "torch>=2.1",
    "lifelines>=0.27",
    "mlflow>=2.9",
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "python-multipart>=0.0.9",
    "pydantic>=2.4",
    "onnx>=1.15",
    "onnxruntime>=1.16",
    "skl2onnx>=1.16",
    "evidently>=0.4",
    "pandera>=0.18",
    "pyarrow>=14.0",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.10",
    "ruff>=0.1",
    "pre-commit>=3.5",
    "httpx>=0.25",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["data*", "features*", "labels*", "evaluation*", "models*", "mlops*", "service*", "monitoring*", "policy*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501", "N999"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
